+++ python/shotgun_model/data_handler_nav.py
@@ -1,0 +1,25 @@
+# Copyright (c) 2016 Shotgun Software Inc.
+#
+# CONFIDENTIAL AND PROPRIETARY
+#
+# This work is provided "AS IS" and subject to the Shotgun Pipeline Toolkit
+# Source Code License included in this distribution package. See LICENSE.
+# By accessing, using, copying or modifying this work you indicate your
+# agreement to the Shotgun Pipeline Toolkit Source Code License. All rights
+# not expressly granted therein are reserved by Shotgun Software Inc.
+
+from .data_handler import ShotgunDataItem, ShotgunDataHandler, log_timing, ShotgunModelDataError
+
+class ShotgunNavDataHandler(ShotgunDataHandler):
+    """
+    Data storage for navigation tree data via the nav_expand API endpoint.
+    """
+
+    def __init__(self, cache_path, parent):
+        """
+        :param cache_path: Path to cache file location
+        :param parent: Parent QT object
+        """
+        super(ShotgunNavDataHandler, self).__init__(cache_path, parent)
+
+
